+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Added NavigationType.custom and `Widget Function(BuildContext, NavigationScreen, int)?customNavigationBuilder` for building it.
   * Along with `NavTab` you can now use `NavDivider` or just `NavWidget` to add custom widgets to the navigation. Currently only supported by drawers, rails, & sidebars.
   * DOC Updated `Screens` tab for new nav tabs (hiding custom)
+* Updated Dependencies & Promoted UUID
 
 ## 3.7.6
 * FIX failed compilation on master